chore(AutoComplete): extract into whole separate component
